requirements.txt
<<<<<<< HEAD
asgiref==3.8.1
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cryptography==41.0.7
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.7
Django==4.2
django-allauth==64.1.0
django-annoying==0.10.6
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-extensions==3.0.2
django-filter==24.3
django-guardian==2.4.0
django-htmx==1.23.2
django-jsonstore==0.5.1
django-modeltranslation==0.19.16
django-viewflow==1.11.0
djangorestframework==3.14.0
filelock==3.12.4
fonttools==4.55.3
graphviz==0.14
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
mysql-connector-python==8.0.20
numpy==1.25.2
oauthlib==3.2.2
pandas==2.1.0
Pillow==10.0.1
pipenv==2023.9.8
platformdirs==3.10.0
prettytable==3.8.0
protobuf==3.12.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pydot==1.4.1
pydyf==0.11.0
pyfiglet==0.8.post1
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==2.4.7
pyphen==0.17.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2020.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
six==1.15.0
sqlparse==0.3.1
tabulate==0.9.0
termcolor==2.3.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2023.3
urllib3==2.1.0
validate-email==1.3
virtualenv==20.24.5
wcwidth==0.2.6
weasyprint==63.1
webencodings==0.5.1
websockets==12.0
zopfli==0.2.3.post1
=======
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
bleach==4.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-summernote==0.8.20.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
html5lib==1.1
idna==3.10
lxml==5.3.1
oscrypto==1.3.0
packaging==25.0
pillow==11.1.0
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pypdf==5.4.0
python-bidi==0.6.6
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
xhtml2pdf==0.2.17
>>>>>>> 1ec4ac0a13ca208754d746c157c9fc6d929e54f6
